+++ CHANGES.md
@@ -1,3 +1,3 @@
 ### Next
- - (Change)
+ - (Feature) Enables support for custom element selectors.